services/archiver/requirements.txt
# Matrix client
matrix-nio[e2e]==0.24.0

# Supabase
supabase==2.3.0

# Web framework for health endpoints
fastapi==0.109.0
uvicorn==0.27.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP client for media downloads
aiohttp==3.9.1

# Async PostgreSQL driver
asyncpg==0.29.0

# Logging
structlog==23.2.0

# Type hints
typing-extensions==4.9.0